+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add Cracked Sand and Sandy Dirt to Sand Block and item tags
 - Make Cypress Saplings use 2 by 2
 - Fix Yucca Leaves Models
+- Add Light to Fairy Slipper
 
 # 1.0.5-Beta
 - Workaround for a crash involving fruit leaves decay (No Drops)